[WIP] CA-183553, CA-187022: split domain creation into several transactions
Currently, xenstored treats a transaction that writes to both /a and /b as
conflicting with a write to /c. Hence if something else writes to /c at high
frequency, it can make the transaction repeatedly fail with EAGAIN and retry
forever.

Xenopsd's domain creation code had a transaction that writes to {/local/domain,
/vm, /vss} so conflicts with a write to any location in xenstore. In normal
operation the regular guest-agent writes cause this transaction to be retried
many times, leading to poor performance (CA-183553). In certain scenarios this
makes it easy for a VM to cause a denial-of-service attack on VM.start
operations (CA-187022).

This patch works around xenstored's over-conservative notion of conflicts by
splitting the domain-creation transaction into three independent transactions,
executed sequentially. (Clearly, having them all together in one transaction is
preferable, so this is undesirable, but is the simplest work-around until
xenstored's conflict model can be improved.)

When splitting these transactions up, we must be sure that nothing else expects
these operations to happen atomically. Specifically, nothing must read/watch
state written by one transaction and assume that the contents of the subsequent
transaction(s) will also be present. We need to be sure of that before
committing this patch.

Signed-off-by: Jonathan Davies <jonathan.davies@citrix.com>