CA-197809: trivial xenstore cache
Whenever we get a xenstore watch on a domain's tree firing, it results in a call
to VM.get_state. This is an expensive call which causes a huge number of
xenstore operations (360 in the case of a simple VM I tested with), most of
which are reads.

Instead, we can maintain a cache of the xenstore contents and update only the
entry relating to the watch that fired, safe in the knowledge that all the other
data is unchanged.

Signed-off-by: Jonathan Davies <jonathan.davies@citrix.com>